Add KeyguardTransitionRepository
Added in order to better control transitions within
keyguard. Transitions must begin with a call to startTransition and
can be either animated or manually controlled. The domain layer is
responsible for both triggering these transitions and listening to
flowable events emitted by this repository, and should respond by
updating the UI elements.

Bug: 195430376
Test: atest KeyguardTransitionRepositoryTest ClockEventControllerTest
Change-Id: I56bd9d560db265c523fa3e0b8c03310e9f4067de

Former-commit-id: 0e8f2a2efff008d93d01f8f5bfbf951a78529918