Let IMMS call isKeyguardSecure() for the correct user
InputMethodManagerService has been using

  KeyguardManager#isKeyguardSecure()

to fine tune when the IME swicher icon should be shown on the lock
screen [1].  However a recent CL made that API @UserHandleAware [2],
which was not reflected in InputMethodManagerService yet.

With this CL, InputMethodManagerService can query isKeyguardSecure()
for the correct user.  A bonus is that this CL removes the direct
dependency on KeyguardManager from InputMethodManagerService.

 [1]: I2ae7fc7676e5b64d8910dbc6833d3042e8d2329a
      436db799860bf4bcb05d4a712806f61706b98712 [formerly d2bc309f0680c72731cf79f1abb300daebf1ec92]
 [2]: Ieb7217f5e1dddb65f8987242e675c3dff0a80aab
      ad525c554aa66a5590db2e93c2e18419deff99fa [formerly 2bdc21e6a996ae1287c2151ac2353146b0815401]

Fix: 251459998
Test: presubmit
Change-Id: Ia29e303fe07fd876ea2ecfb937c1fb5b867a0e89

Former-commit-id: e6540d2bcb85740083aeed40a07726a8b6703e37