Fix BiometricPrompt retrying without finger lift
Authentication errors from BiometricPrompt would result in a call to
`onCancelUdfps` to turn off the UDFPS illumination. As a side effect,
`onCancelUdfps` would clear the internal state machine, including the
fact that a finger is touching the sensor. Consequently, any natural
micromovement of the finger would register as a new touch and
lead to a new authentication attempt.

We no longer need `onCancelUdfps` to turn off the UDFPS illumination
early. This is handled in `UdfpsOverlayController#onAcquired` and
the lifecycle methods of the fingerprint client monitors.

Bug: 236338150
Test: atest SystemUITests:com.android.systemui.biometrics
Change-Id: I5110a798e9557a7df55f42eac467a3da4c2d4c00

Former-commit-id: 2079d344c1b8f5ac46e3ac9d85522d91157c175d