Don't always suppres redraw report request in relayout
I6faad4b9adac7334df7caf3d23df59bd3e96bfd9 added an optimization so a
newly visible window won't be relayouted several times with shell
transitions but this seems to conflict with
Ie545465cac0207631ada6dac003340ca80eb2b14 that the relayout won't
return syncId to trigger reporting finish draw.

Especially when a window size gets changed but visibility doesn't, we
seem to rely on the code path of relayout->reportResize. If this code
path is suppressed, a new buffer won't get drawn for the change and the
transition will time out.

Bug: 237743987
Test: Multi-activity PIP works on ARC T
Change-Id: Id2b5a335c05f5977a7b72e58470218f075c302f9

Former-commit-id: 1a721d6022d280e912a372759ef941f0314b3a0f