Suppress scrolling when handwriting initiator requests focus
When stylus handwriting is started on an unfocused view which is
partially scrolled off screen, the handwriting initiator will request
focus on the view, which can cause the scroll position to jump to move
the view to the visible area. This is particularly problematic if
performing a rich gesture (such as circling to select some text), since
the position of the gesture relative to the view is important.

View#setRevealOnFocusHint is used to suppress the scroll. This setting
is currently respected by ScrollView and HorizontalScrollView.

Bug: 250920261
Test: Manually tested
Change-Id: I0da67f48876ab52e92a463992e8075c983e71819

Former-commit-id: 92c48e3e03c9947342fa3af31be2090f7346c95c