Move the ActivityEmbedding Task visibility filter to the client
Before, we only send TaskFragment info event to the client side when the
Task is visible, which may result some TaskFragments stuck in an
intermediate state. Now, we also send the info event to the client if
the app owns any visible Task, and let the client organizer to determine
if it wants to update the TaskFragments.

Bug: 251583070
Fix: 250943776
Fix: 251132298
Test: atest WMJetpackUnitTests:SplitControllerTest
Test: atest WMJetpackUnitTests:TaskFragmentContainerTest
Test: atest WmTests:TaskFragmentOrganizerControllerTest
Change-Id: I47570639c4ed4bdfb4202f65ba4531adbd4a2d2d

Former-commit-id: 5dafb54f14fa66857b9453bc6fb8e8e1386ef8b1