Fix concurrency issue in PropertyInvalidatedCache
Bug: 253063488

Ensure the global lock is held while fetching the list of active
caches.  The global lock is necessary while the list is being fetched;
it is not necessary while each cache in the list is cleared.

Added a new test for PropertyInvalidatedCache.onTrimMemory().  This
test will not catch race conditions but does verify that
onTrimMemory() behaves as expected.

Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
Change-Id: I5022620cd4f2561179af709246a9bf149423143f

Former-commit-id: 803254d0ab3109165ffc1e3f409c07a692a9ac37