Simplify FINISH_INPUT_NO_FALLBACK_CONNECTION handling
This is the 6th CL for Bug 26851566 [1][2][3][4][5].

In the CLs [3][5], the approach we took was to extend an existing IPC
method

  IInputMethodClient#setActive(boolean active, boolean fullscreen)

by adding one more parameter "reportToImeController".

However, it seems that such an approach made its handling code hard to
maintain, because the IPC method needed to behave in a completely
different manner depending on the newly introduced parameter.

To avoid such a confusion and to make subsequent refactoring easier,
this CL introduces a new IPC method

  IInputMethodClient#setInteractive(
          boolean interactive, boolean fullscreen)

to clarify what we are actually doing.

In most cases there must be no observable behavior change.  One
possible behavior change is when setInteractive() gets called when
IMM#mCurRootView.getView() is not available.  In the previous CL, we
fell back to previous behavior as if
FINISH_INPUT_NO_FALLBACK_CONNECTION was not defined.  However it looks
like such a fallback mechanism is not an intentional behavior.

The main use case is still verified with existing CTS tests [6][7].

 [1]: Ic11956fe745a829da1ec7c09ea214e9b3961c8e0
      89136b7a695f4e0ddc60e28e44ffd1ed492d88a6 [formerly 0df8812486ff0b104622dec72d6243a783b10257]
 [2]: If06daf71160aa44a4254ac125561974ecbdef4f2
      8009cae89caa9cb002209e31dde148088462d630 [formerly 2c6e80be1253140f1761c1bb4cc601e7e350e1b2]
 [3]: I8a657e75e274d842fb46b60375f6aeafeab96a59
      7b1c002f0761c94edb9fc0e0ccc2fbf2eaa11640 [formerly 152eacec4f5e1c78eafd42abbcf62fc12b0eee33]
 [4]: I16f4a34360a2f64b69978724648a9be741f140b5
      03874f380ef6afc6c17ece76f55ccda4339a662c [formerly 3d62f739a869c97a265506ac1e1ab58f93fa3b34]
 [5]: Id4e71a822bfde5fe6a263bbe094c0d238017efe1
      03874f380ef6afc6c17ece76f55ccda4339a662c [formerly 3d62f739a869c97a265506ac1e1ab58f93fa3b34]
 [6]: Iba0332ed3b767a9a15768b64947b9cde31105af1
      e2f4138156c784746becde6a01c19f0c3de6ac32
 [7]: I346ef42f2546a2925bf44ca7719ae19399e15e78
      84c5ec2f81d4be5285b9c5f1d0e943f808e9158c

Bug: 26851566
Bug: 156215187
Test: atest InputMethodStartInputLifecycleTest
Change-Id: I0c95d7e70357ac98b336cd1b51a7405ba92a9a6f

Former-commit-id: 3ca950724ccb01a0371ba16ffd8ee74affcb365f