Merge WINDOW_FOCUS_GAIN_REPORT_WITH{,OUT}_CONNECTION again
This is the 5th CL after our CLs [1][2][3][4] for IME visibility
issues.

While the last CL [4] logically reverted most of changes made in
previous CLs [1][2][3], there remain some unintentional changes around
WINDOW_FOCUS_GAIN*.  The biggest issue is the invocation of

  IMM#mServedInputConnectionWrapper.isActive(),

from startInputOnWindowFocusGain() added in the 3rd CL [3].  This is
really problematic because it internally

  IMM#checkFocus(),

which is not intended to be called from startInputOnWindowFocusGain().

This CL addresses that problem by effectively removing the call path.

App compat risk is supposed to be low because IMM#checkFocus() was not
called before the 3rd CL [3], which was introduced in Android R.

There are CTS tests to verify that related issues are not regressed
again [5][6][7].

 [1]: I45a9814d812ad906f417c24200fd4219959e2423
      04b5f52fa1aa5bc82cd0cca2a37a244b3bc4b884 [formerly 6ecbf08772d73cea6b78b5d8eae4e925e0599385]
 [2]: I2da99ae67b9ce4051dec0c0f0e975ebe6e1ab118
      2f05138e371f0d25cca2b4543c26d0562f299b37 [formerly 48bfc3165c4c688ee5a8e78cf6d2b685f1ad032b]
 [3]: I1ef3d341af9d473d94d52fd1890deafbae2bc9e1
      993817f9584d51988672a09eef186325da03f674 [formerly d43b75cf0edea197ed2ee690814eedf6844aadbe]
 [4]: I022ad658844142ff4a4cf3b91953013f2bfbb58a
      95db463a454e68af790c329ee0c9977c6135745a [formerly a24622cfcdb341d62f0c65137361b6d5f9220151]
 [5]: Ie029c109e257be81a1ad6327a706bb6bc1e8316c
      3045df602f061a185f6fff18e81f5af1a2c51270
 [6]: Id48eedc1a2d5597bbba5f1d3fc01540e2f6de4c5
      3ed79ab2de600bce312b4ce59b319da0ad9a8dac
 [7]: I022ad658844142ff4a4cf3b91953013f2bfbb58a
      95db463a454e68af790c329ee0c9977c6135745a [formerly a24622cfcdb341d62f0c65137361b6d5f9220151]

Bug: 152373385
Bug: 155781821
Bug: 160391516
Bug: 161273049
Bug: 182259171
Test: presubmit
Change-Id: I64412e87c4881018ea67140f9472e10ea41836e4

Former-commit-id: 62f93a5cf9f6dcc222cf218b7a88460abf90b5d5