Handwriting initiation delegation
Handwriting initiation delegation is introduced to support views which
appear to be a text input field, but are actually non-editor views which
are linked to a real editor view. Such views can be configured as
delegates, so that stylus movement on the view can trigger the
transition to show, focus, and start handwriting on the real editor
view.

Bug: 228100684
Test: atest android.view.stylus.HandwritingInitiatorTest
Test: atest android.view.inputmethod.cts.StylusHandwritingTest
Change-Id: I492fd507895d59213753c4ad06d42a4a212e0fde

Former-commit-id: 9299b0b4efec5abd0513a7c93f00327d74711ac8