Lazy UserSwitcherDialogCoordinator.
Makes all class dependencies lazy and turns on the refactored code.

The motivation was the attached bug, where we've seen a regression in
boot times when turning on the refactored code flags. The theory is that
this core startable is pulling in a lot of code that's doing work in its
constructor(s). By making it lazy, that's delayed until the core
startable system tells this class to start.

Bug: 253435919
Test: Manually verified that the user switcher works.
Change-Id: I8957d9a98ca8e949f888b84c9d0d58d51f15f9d5

Former-commit-id: 123d7cb64d99d34a5cf66e9205b72f6a50a96155