Add IS_REMOTE extra for ClipData
This extra can be used to indicate that a ClipData comes from a
separate device rather than being local. It is a rendering hint
that can be used to take different behavior based on the source
device of copied data.

Bug: 246808749
Test: manual (no behavior change)
Merged-In: I6998b7a216fac73a1387367f4179efb95d7562c5
Change-Id: I6998b7a216fac73a1387367f4179efb95d7562c5

Former-commit-id: b4a5acfcc2a13fc1968f5e6296ac77c410247394