Make IMM#hasActiveInputConnection() more predictable
This is a follow up CL to our previous CL [1], which introduced

  InputMethodManager#hasActiveInputConnection(View)

as a test API.

As this remains to be a test API, we should be able to tweak its
behavior as long as existing use cases continue working.

The main motivations of this CL is to make it more predictable and
easier to understand.

Even with simplification, ImeAwareEditText and CTS tests should
continue working as expected.  Also there must be no app compatibility
concern in this CL.

 [1]: I466a93e2b4706ae809ac50dc16dc8e4ffa7c2f1c
      2542e2ea19b5adbed90aad3dd8dca369f31a3c61 [formerly 8bcce88c8608d77a2fb760e932fed8f08e3645fe]

Bug: 176697749
Bug: 182259171
Test: presubmit
Change-Id: I7fe8b18ffbb450fd1e272e15f06f2629518f320c

Former-commit-id: bf5dc9ecbf0f2bb8bb04e367de049d13e89759dc