Allow upcoming vibration when current one if being cancelled
The vibration cancellation at VibratorManagerService is asynchronous and
might take some time before it sets the current Vibration status to the
cancelled one requested. If a new binder call to vibrate arrived before
the status is updated then the new vibration might be ignored for the
current one, even though it's already being cancelled.

Add a check to the cancel signal sent to the VibrationStepConductor
before deciding to ignore new vibrations in favor of the ongoing.

Bug: 249032223
Test: VibratorManagerServiceTest
Change-Id: I8151d9ee0e30b7115e9cb3d6bcd272da2806b45c

Former-commit-id: 546dd659054de251ff64f4b3b4c9c72f0fe514d4