Implement DISALLOW_ULTRA_WIDEBAND_RADIO AE user restriction
Ultra-wideband (UWB) is a radio technology that can use a very low
energy level for short-range, high-bandwidth communications over a
large portion of the radio spectrum. This CL adds a user restriction
for disallowing UWB on a device. A user restriction is needed for
NIAP (Common Criteria) certification (see b/239714441). One of the
requirements for the certification is to be able to enable/disable
all the radios.

Bug: 239714441
Test: m; manual testing with testDPC;
Change-Id: Icfa16258cc789c337417c2e96a4e0c37d0030d55

Former-commit-id: 6ee8817cf7eb91b1df5bef7a85c7bdb567450559