Expose PackageState#getSeInfo
Exposed through PackageState instead of AndroidPackage so that the
overrideSeInfo can be applied through PackageState instead of
exposing the value separately and relying on the caller to apply
the override.

This also moves the backing field into PackageSetting, since
there's no reason to keep it in PackageImpl now. It's always
re-set on package scan.

Bug: 254029037

Change-Id: I703a45a86d9c7b4e27f225b3cf2c66427744c32a

Former-commit-id: 6e67aa16bd630465c8494f889c3ec54637b9aba3