[Bouncer] Fix MODERN BOUNCER issues.
Fixes a CTS failure by ensuring that isShowing is false when we are invoking our disappear animation finish runnable. We do this buy setting the state flow to null after we run the runnable.

Fix an instance where we are calling bouncer.isShowing and transition it to the correct method.

Fix BouncerView implementation. We set the delegate after the constructor of the StatusBarkeyguardViewManager so holding it in reference makes it null.

Bug: 240299776
Test: Manual and  atest KeyguardLockedTests#testEnterPipOverKeyguard --verbose

Change-Id: I62f24e06481fcabfcfe6f5d5469b2513fef0ec83

Former-commit-id: 8ed0d525afc9efb076d337ffe3b10b9336f49bba