Don't prepare CE storage on user creation
Since the CE key is now always protected by the synthetic password, it
is no longer saved to disk before PHASE_BOOT_COMPLETED.  Unfortunately,
that means that if the device is rebooted after a user was created but
before the CE key was written to disk, then the CE key needs to be
regenerated.  But that doesn't work if CE storage was already prepared.

Address this by not preparing the user's CE storage at creation time.
It has never actually been necessary, since the CE storage will be
prepared when the user is unlocked anyway.  This is also how user 0
works already (with the exception of the /data/data directory).

Bug: 232452368
Bug: 251213447
Change-Id: I78ba39030bbd3d1561f822fc7711d7169a21c467

Former-commit-id: 5c65b1ee1023c7d9aa449cfcfb22830afac0900f