Sort running task by focus and visibility
Currently the order of visible tasks we get is opposite to that in the
system.

This CL is to ensure that the visible focused task is on top, visible
tasks below, then invisible tasks. And keep the order within the
acquired focus, visible and invisible tasks groups consistent with
the system instead of sort by lastActiveTime.

If there are multiple DisplayContents, getTasks method will return:
[{VisibleFocusTask for DisplayTop, VisibleFocusTask for DisplayBottom},
{VisbleTask0...n1 for DisplayTop, VisbleTask0...n2 for DisplayBottom},
{InvisbleTask0...n3 for DisplayTop, InvisbleTask0...n4 for DisplayBottom}]
in the same result List.

Bug: 255255158

Signed-off-by: Bowen Li <libowen1@xiaomi.corp-partner.google.com>
Merged-In: I7f7d5e72fbf73988b58c23f8a11dff05dbc1b3f1
Change-Id: I7f7d5e72fbf73988b58c23f8a11dff05dbc1b3f1

Former-commit-id: 76df3f45740e10f78aac3d406ff9b81cef6249a8