Change dream IN animation duration to 250ms.
This change adds a framework level config for dream IN/OUT animation
durations that is shared by the animation xml and keyguard view mediator
to ensure wherever the dream is launched, it has a consistent launch
animation duration.

The dream IN duration has been changed to 250ms from previously 1000ms
to match specs.

Bug: 222507937
Test: observe dream IN animation over launcher
Test: observe dream IN animation over keyguard
Change-Id: I5a51406f8476a76131b6c1aeeb89d12bd43a5a17

Former-commit-id: 58adc76595d1e66c1fe4d9a071f1ade25392654d