Revert "Enforce ComponentName belongs to caller app"
This reverts commit bfe5dad23330c86f9810b224a9aa2a0748308d4d [formerly 7319e40b37aba598e081b56095b7ed4e0a04396c].

This revert was manually created from the patch in ag/19748522,
but is not a revert created using the gerrit UI because of
b/251778962.

Reason for revert: b/241798893. Note: this fix is not the root
cause of the breakage. It just exposes a previously unknown bug.
See linked bug.

Bug: 251778962
Bug: 241798893
Bug: 238177121
Test: atest CtsMediaBetterTogetherTestCases
Change-Id: I01bda0f5bf343d73c480ad64b562dd0a2ee46e30
Merged-In: I01bda0f5bf343d73c480ad64b562dd0a2ee46e30

Former-commit-id: 1597cde5790126b7a899f08447681701896f1b99