HDMICEC: Initialize HdmiControlService with local devices in unit tests
Initialize HdmiControlService with local devices in unit tests and
remove uncessary allocateLogicalAddress calls.

The device types list is set in HdmiControlService's constructor. With the help of this list, corresponding local devices are allocated whenever initializeCec is called (f.e. HdmiControlService#initService)  such that extra calls of allocateLogicalAddress can be avoided.

Bug: 183081487
Test: atest com.android.server.hdmi
Change-Id: I750a6d937cab552b6ed6f98dfa2201fa86ba5aa4

Former-commit-id: 6eb366060e281c7165ad76611e0f8ce68c8cb71f