Moves quick affordances to data layer.
Because we need to implement a data layer repository that will access
quick affordance configs, we need the latter to be moved to the data
layer as well (since currently they are in the domain layer and code in
the data layer is not allowed to depend on code in the domain layer).

This CL has no changes, only moves and renames.

Bug: 254858696
Test: All tests still pass. Bottom area affordances still work.
Change-Id: Ibe1f0387886950ef7e9ef11a8716603aa547493e

Former-commit-id: 7237112d22d8412bbbcea5ebe43dbfd383f14bc3