[SB refactor] Track the default subscription and its connection status
This CL adds the ability for `MobileIconsInteractor` to track which
`MobileConnectionRepository` represents the active data subscription
connection, and tracks its disabled status.

In the case of DSDS, for instance, data will not be enabled on the
stanby SIM, but we will only want to show the disconnected state (the
exclamation mark on the triangle) when the active mobile data
subscription has data turned off.

Also, added `ContentObserver`s for the `MOBILE_DATA` setting, which is
used to trigger telephony polling events. This mimics behavior from the
old `MobileSignalController` and help us to update the icon once the
data enabled state changes.

Test: atest MobileConnectionsRepositoryTest
MobileConnectionRepositoryTest MobileIconsInteractorTest
MobileIconInteractorTest MobileIconViewModelTest

Bug: 240492102

Change-Id: I6f614d615930dca1d0c69ef361ecc05ce394dbfc

Former-commit-id: be25e8f1f407df340636259f6d22496638e8b507