Revert "libandroidfw: use map for more efficient indexOfString()"
This reverts commit 1c5520c1d6b9579f069b337eebde0a49dadc0bd0 [formerly 63e5d41be79f02cf827f395419a93fb8e37f8a47].

Reason for revert: b/253420843

This CL causes memory regression as described in above bug. To speed
up the lookup from resource string to resource ID, the original CL
introduced std::unordered_map for the purpose, unfortunately, that
adds much memory footprint due to the number of key strings per
process. We will need to revisit the solution with a more memory
efficient approach.

Change-Id: I9c6754d361f091e963d4bebea0e698e72e221e51

Former-commit-id: f9f51c819786d8ef231e9eb0ec28688a5575366d