Exclude unnecessary animations of shell transition
There may have multiple transitions when launching a task with
embedded activities. In the last transition info, it may contain
activities occluded by starting window and a closing wallpaper.
By default, all of them will be animated with edge extension,
that looks like showing some noise.

Because the animation of embedded activities under a task level
starting window is not visible, it can be skipped. But for
non-embedded activity, the activity level starting window can be
changed or closed with the host activity, so the case should not
be skipped.

Besides by default, wallpaper is no animation, especially it
shouldn't apply any task/activity style animation.

Also make sure the leash of starting window is always on top of
the task with embedded activities. This just makes the surface
hierarchy consistent.

Bug: 255269113
Test: No flickering when cold launch Settings on a large screen
      device with support of activity embedded.
Change-Id: I8fd1137e806aeb4060a83a71dc5aa02acdc42429

Former-commit-id: e65d74270d0098829735e44b21bb3203dc49bb60