Don’t allow activities with different display categories in same task
Activities now can be declared to be displayed only on a particular
display category (like Auto or AR), Activities of different display
categories should not be placed on the same Task.

This CL considers the new attribute of activity while looking for
the existing task during the activity launch. The new_task flag is
added if needed to prevent launch in the source record’s task. Also,
prevent the activity landing on a Task of another display category
if mInTask is set.

Bug: 253324111
Test: atest ActivityStarterTests
Change-Id: I9da35a0c483303bc1e15cc05fb50e384c943dcca

Former-commit-id: 5ef83c770c155ce4268c851a78aa32cb4bd95389