[Unfold animation] Do not send start closing event before the inner display turned on
Currently, if the hinge angle sensor reported
a decreasing value before the inner screen
fully turned on, we can see a flicker because
it starts the fold animation (from progress=1.0f).
It is pretty consistently reproducible when
unfolding only to half foleded posture.

Updated the DeviceFoldStateProvider to not
report start closing event if the screen
hasn't been turned on.

Bug: 233045200
Test: atest com.android.systemui.unfold.updates.DeviceFoldStateProviderTest
Test: unfold to half-folded state multiple times
  => verify that there is no flicker
Change-Id: I516dbad67d201ead03e3dae593f12fc5a985447b

Former-commit-id: 244d112adb2bfe0f4966c809c312674ab8331583