Add logic to show apps running certain jobs in Task Manager.
For now, we will show apps with an active FGS or jobs that the user
should be aware of (ie. "user-visible jobs") in the exact same way.
Users will be able to stop the ongoing FGS and user-visible jobs
using the same button. This won't be used to show apps on TM, so
the JobScheduler implementation is empty.
We'll add the JobScheduler implementation in U.

Bug: 255767350
Change-Id: Iaf482926c1f568aee26a2798ce827c6138f3f779
Test: atest SystemUITests
Test: Manually start jobs and interact with the Task Manager

Former-commit-id: 4167f0c3d6e0543f04029a44673080a48c925937