[SB refactor] Upgrade most data sources to `StateFlow`
Most of the data model for mobile is more correctly represented as a
StateFlow, since the state from callbacks needs to persist even when
there is no corrent collector on the mobile icon.

This change also exposes a current issue: toggling mobile data on/off
multiple times can introduce an inconsistency in the UI. This is
approximately the same as the old pipeline, which also experiences the
same problem.

Test: manual
Test: atest packages/SystemUI/tests/src/com/android/systemui/statusbar/pipeline/mobile/*
Bug: 240492102
Change-Id: Idd5fa0205380b4ba15f7d723b32b08e114456f7f

Former-commit-id: 93b945ce54f1e61d75e3cf9cef4a4083f03c5e24