UserVisibilityMediator refactoring, step 1.
The initial version of UserVisibilityMediator was relying on
UserManagerService to get the state of the current users (full and
profiles), which caused a cyclic dependency between these 2 objects.

This CL breaks that dependency by replicating such state into
UserVisibilityMediator (and adding methods to start / stop users).

There's still room for improvement (just look at the TODOs :-), but
they'll be handled on future CLs, so this one minimizes the changes
(both in the main code and unit tests).

Test: atest FrameworksMockingServicesTests:com.android.server.pm.UserManagerServiceTest UserVisibilityMediatorMUMDTest UserVisibilityMediatorSUSDTest UserControllerTest
Test: atest CtsMultiUserTestCases:android.multiuser.cts.MultipleUsersOnMultipleDisplaysTest
Test: adb shell dumpsys user --visibility-mediator

Bug: 244644281

Change-Id: I8daa751dbc6fd4c38edd21ab9ef7b7c7ad13c8ff

Former-commit-id: a2c65f11be13758d1535e178843e1ad9e2ab7bdd