Fix win mode issues for unresizable apps
Some unresizable apps are launched in freeform in freeform display,
but the override windowing mode of this task needs to be left
UNDEFINED. Otherwise, the task would remain freeform even when the
display transitions from freeform to fullscreen.

On the other hand, fullscreen unresizable apps need to remain
fullscreen regardless of the windowing mode of the display.

Also, with this CL, the same unresizable policies are applied in
fullscreen display too as the display could be freeform later
anyway.

Bug: 185427982
Test: atest WmTests:TaskLaunchParamsModifierTests
Change-Id: I7279be115d88731566c19be691ff562d5885a3af

Former-commit-id: 383b0006c5e024c3dc63bc0884a7aae23c039516