Fix guest user not being removed until boot
Updated the GuestUserInteractor on user switcher to use
removeUserWhenPossible() instead of removeUser(), similar to what
happens on UserSettings.

This avoids the case where the guest user markes itself to be deleted,
but cannot be removed by the UserManager because it is the current user,
so it will be purged only when the device reboots. With
removeUserWhenPossible(), the guest will be remove uppon reboot or user
switch.

Bug: 250544682
Test: atest GuestUserInteractorTest &&  manually tested
Change-Id: I3b185842348e452b18da4cd346062656a1043114

Former-commit-id: 1cc5912bd2419d09a812bed6e800a34c6b6bbe16