Make SurfaceSyncGroup AIDL interface and add cross process syncs
When two processes are involved in a sync, we don't want them to
exchange Transaction info because it may contain SurfaceControls that
the two transactions shouldn't exchange. Instead, have system server
take care of the cross process syncs and ensure SurfaceSyncGroup
sends the data to system server instead of to the two apps.

This is done by the following flow:
1. P1 creates a SyncGroup
2. P1 adds something local to the SyncGroup. No WM API is called.
3. P1 adds something in P2 to SyncGroup
3a. P1 creates a SyncGroup in WM with P1's SyncGroup as token
3b. P1 notifies P2 that it should be added to the SyncGroup with a
specified token
3c. P2 invokes WM call and adds itself to the same SyncGroup in WM

When P1 and P2 groups are done, they will invoke the transaction
callback to WM, who will merge into a single transaction

4. P1 marks SyncGroup as ready, which calls into WM to mark the
   associated SyncGroup as ready. When all callbacks are complete, it
   will apply the final transaction

Test: SurfaceSyncGroupContinuousTest
Bug: 237804605
Change-Id: I0f52f2e2d1a95e0ecc2b95b6dddd2ebb5b3edb54

Former-commit-id: b8afa7f7b99936e4313608965765dfaf6863cce2