Content provider for quick affordances.
This is the content provider that wallpaper picker will be using to
query for slots, affordances, and selections, and for setting or
unsetting selections.

Fix: 254857637
Test: Manually verified the API by
issuing adb shell content commands and seeing affordances appear and
disappear on the actual lock screen.

Change-Id: Ia47a8ee322028042bb3e118ddb82a92a27eae191
Merged-In: Ia47a8ee322028042bb3e118ddb82a92a27eae191

Former-commit-id: 04777dbab2221d202b0f4c05bb6d3fcaa9ca625e