Make TEST_INPUT_METHOD @TestAPI
This is a follow up CL to my previous CL [1], which introduced

  android.permission.TEST_INPUT_METHO

to protect several @TestApi methods in InputMethodManager.

This CL makes the permission itself as @TestApi so that it can be
included as an API constant in test-current.txt.  This also enables us
to minimize special permission while running CTS as follows.

  SystemUtil.runWithShellPermissionIdentity(() -> {
      // do something.
  }, Manifest.permission.TEST_INPUT_METHOD);

There must be no user observable behavior change in this CL.

 [1]: Ie79a3e9d41ce22605ae083594d639c37d08b7def
      e67074ad51444e3a7df90523ecf01246bd8e8f35 [formerly b869c783808c4d4937727a1672b2fac81bace368]

Bug: 256239860
Test: presubmit
Change-Id: I1fd3ceda9a1b641e63c7633a223a9a806e78cd2d

Former-commit-id: 89d683f2cfec03b1f244d9b64c151de893d4daf5