Require TEST_INPUT_METHOD instead in IMM#addVirtualStylusIdForTestSession()
This is a follow up CL to our previous CL [1], which introduced an @TestApi

  InputMethodManager#addVirtualStylusIdForTestSession(),

which internally requires INJECT_EVENTS permission.

Now that TEST_INPUT_METHOD is also available [2], let's require it
instead of INJECT_EVENTS for better consistency with other test APIs
defined in InputMethodManager.

This CL also add @RequiresPermission(TEST_INPUT_METHOD) to

  InputMethodManager#addVirtualStylusIdForTestSession()

to make it clear that it's guarded with the permission.

 [1]: I7399c0a56c04f024d65c16d459818cc2c5cdbc7f
      83b9523b735b56f97785b1c949dbb44ddde3ca2b [formerly d96ef57bc4f87f82395c4eb456b4c11c6faf912b]
 [2]: Ie79a3e9d41ce22605ae083594d639c37d08b7def
      e67074ad51444e3a7df90523ecf01246bd8e8f35 [formerly b869c783808c4d4937727a1672b2fac81bace368]

Bug: 256239860
Test: atest CtsInputMethodTestCases:StylusHandwritingTest
Change-Id: Icdbb4ade2bd4b5066984e209efcf0ca93bfa08ca

Former-commit-id: cba800ef2f879a8e4bbd69754dd9a56012e5f842