Add enableOnBackInvokedCallaback manifest attribute for activity
Currently, there is already a flag for an application to opt-in of the
new back navigation system.
This will add a new manifest flag for an activity could opt-out/opt-in
of the new behavior.

When the attribute has set, it will add a private flag to ActivityInfo
to identify current activity will opt-in or opt-out the new back
navigation system.

If no attribute set for activity, it will still check the attribute of
application info.

Test: atest BackNavigationControllerTests
Test: atest BackNavigationLegacyGestureTest BackNavigationLegacyTest
Test: atest BackNavigationTests
Test: atest WindowOnBackInvokedDispatcherTest BackNavigationTest
Test: atset BackAnimationControllerTest
Bug: 256536263
Change-Id: Ic126c0d4ce54d146698766154e22a18ccb0042eb

Former-commit-id: 032e267622baea135268abf48ed19a54a1b5754c