Make sure userSwitchComplete is dispatched after screen is unfrozen.
At the end of user switch flow, registered UserSwitchObservers are
informed that user switch is complete. This should be after screen
is unfrozen. But if there is no password on the target user, keyguard
is dismissed just before unfreezing the screen and unfreezing is
postponed in handler. This causes the order to be as follows:
dismissKeyguard -> dispatchUserSwitchComplete -> unfreezeScreen

This CL makes sure the order is as follows:
dismissKeyguard -> unfreezeScreen -> dispatchUserSwitchComplete

Bug: 256576977
Test: atest FrameworksServicesTests:UserControllerTest
Change-Id: Iada3131b611526a8c7a372b4500bcbc17299dcbc

Former-commit-id: 7fb0d1dcb013899a7a30799e2df6c6b6cebf317e