Updates location services to support concurrent multi user.
Allows users that run on multi display to control MLS independently by
redefining whether a registration is currently active or not. For an
registration to be considered active, one of the necessary conditions
is the caller being a current user. With this change, the condition is
changed to be the caller being a visible user.

Bug: 236139244
Bug: 241604546

Test: atest FrameworksMockingServicesTests:LocationProviderManagerTest
Test: atest FrameworksMockingServicesTests:SystemUserInfoHelperTest
Test: manual test in emulator

Change-Id: I8cf1d6148a0465fe85953edc92273cc926f7f40a

Former-commit-id: 361e41067015d7aa4475eca50ed71a624ba7ea0a