Fix launching ActivityGroup with ActivityEmbedding
ActivityGroup has been deprecated since Android H, but there are still
apps using it. When it is used, the child Activity will not have its own
corresponding ActivityRecord. If we treat it as normal Activity, there
will be issue because we can't find Task with its token, nor reparent it
with its token.

Skip child Activity (the one launched in ActivityGroup) for any
lifecycle event.

Fix: 256533615
Test: manually verify
Change-Id: Ibccf1575c3b229a0912f843f6468a6c3e24a893f

Former-commit-id: 5a22eceefeeccc61c48925bb11d07405811f825f