User switcher: hooks up broadcast receivers.
This was a missed part of the feature set. When switching to the guest
user, a notification is meant to be set shown. Hooking up the broadcast
receivers into the GuestUserInteractor achieves this and also shows up a
"resume session?" confirmation dialog, which is what the failed test in
the attached bug was looking for.

Fix: 256835620
Test: "Welcome back" dialog shows now when I switch to guest. Also the
notification shows up.

Change-Id: I4548adce783875ba188b1923a2cc30173d3da3c9

Former-commit-id: d41a9d53f612e421b7cbf8611ee067bf77a00074