Clarify required permission in IMM#getInputMethodListAsUser()
This is a follow up CL to our previous CL [1], which exposed

  InputMethodManager#getInputMethodListAsUser(int userId) [2]

as an @TestApi.

One thing to clarify is that INTERACT_ACROSS_USERS_FULL permission is
required when and only when "userId" is different from the one of the
calling process.  To make it clear, this CL adds "conditional = true"
like other @RequiresPermission usages in InputMethodManager.

This is just about (Test) API annotation.  There must be no observable
behavior change.

 [1]: I4b73d9374d7a2ce4c9dfc48682a43d5d48b39042
      3977f2827573384abc60e7fadc584d61a1c97244 [formerly ec8b1449f257337dab95c8e4c4957c749b1291d8]
 [2]: I5b50b5fe143c74c87b331bda3e5bcc4d6248436e
      f69594b61b203f94eb017dc7fdaf2e0f230d6c7d [formerly d20eef826c879c3be3be1d22322e51ee92880561]

Bug: 256239860
Test: presubmit
Change-Id: I14867e528d4292f1f8713b1d9b491fc5ed6c1abb

Former-commit-id: 05e924b2ab80d1b6eb8bd8b7cb4c50669fb3c8af