Animate PiP away on inset changes.
This specifically refers to the taskbar that can be expanded or
minimized, which changes the bottom inset, but PiP remained in it's
original position. This change will move it down if it was already by
the bottom edge, or move it up just enough to be within the movement
bounds.

Bug: 256560961
Test: manually
Test: http://recall/-/ekEuGtt9d9HWqkUtAzpHx8/dSTq2oAg1k85bpIZRBDLIH

Change-Id: I487be6f54b2f9afae8d18df8cf6456858704c703

Former-commit-id: 9621f52c010fc882abae89ddf0c70612241adb1c