Fix splitscreen test: SnapToDismiss
The test cases verify onSnappedToDismiss callbacks properly invoked
when the snap target indicates dismissing split. The callback is
invoked more than once since the start position is the same as the
end position when flingDividePosition is invoked. This CL updates
the current position to make sure the callback is invoked once.

Bug: 257348773
Test: atest SplitLayoutTests
Change-Id: I5742d3db1ab6d5c80c4b4ffa1a9465655c73ab11

Former-commit-id: b44a40442b70649d7d0ddf58a5ed8ccbba0d1fbb