OMS: do not hide reason for failing transactions (on non-user builds)
To improve security, the reason why an overlay transaction failed is
hidden from the caller unless OverlayManagerService.DEBUG is set to
true.

Tweak the logic to not hide the reason on userdebug and eng builds: this
will make it easier to debug tests that fail in CI while keeping the
code secure on user builds.

Test: manual: make executeAllRequests throw, test via `adb shell cmd overlay enable`, check error message
Change-Id: I1c35b9b24a01ea30a50398df9f250ad4e43c0686
Merged-In: I1c35b9b24a01ea30a50398df9f250ad4e43c0686

Former-commit-id: 5a3163d2eb43279f51d3711fbe2d85f25da69314