Remove AndroidPackage#is_ partition APIs
These are moved to PackageState so that they can be queried in a
single location regardless of whether the physical APKs are
available on device (getAndroidPackage would be null in that case).

This also means any existing framework callers have to be migrated
to PackageState, but that ended up being fairly simple for all
methods except isSystem, which has a lot of usages.

That will be addressed in a follow-up.

This CL does not move the physical data storage of these booleans
to PackageState, but that should be done in the future, so that
the underlying fields can be removed fro, PackageImpl.

That will likely have to be done with a full install refactor, as
there's no PackageSetting available at some points where the flags
are queried, and so no holder for the fields without some messy
temporary objects.

This also removes the SettingsFlag set(Private)Flags mask
restrictor, since that legacy subset flag behavior isn't worth
keeping.

API-Coverage-Bug: 262593427

Test: atest AndroidPackageTest
Test: atest OverlayActorEnforcerTests
Test: atest SELinuxMMACTest

Change-Id: I2a915e55f609992729e082815d78fd2d73cb79fd

Former-commit-id: d4c21606ba0374518db0caeada6be521ab32d4b9