Add timeout fallback to UserController dismissKeyguard on user switch.
After a user switch, keyguard is dismissed if there is no password
for the target user. Then after the dismissal of the keyguard,
unfreezing the screen and calling UserSwitchObservers
onUserSwitchComplete is done. In case dismissKeyguard's callbacks
are not called, user switch can not be complete. This CL makes sure
user switch flow moves on even when the keyguard dismissal fails.

Bug: 256576977
Test: atest FrameworksServicesTests:UserControllerTest
Change-Id: I608a43de4d379a26ef66737c1bece7508d476b2e

Former-commit-id: f2f7a74feebd2417dd203db8e7ed7946cfeee109