Keep status bar icons hidden during camera launch.
Before:
https://drive.google.com/file/d/1FmJFP8ZZ0J6r0DWnfD9HR5JeWVEizHI_/view?usp=share_link

After:
https://drive.google.com/file/d/1U743mR-0ZzzJxK-rDOl2rZiFdEio9EDg/view?usp=share_link

This has been a long standing issue with camera launch that
looks pretty bad. When we start occluding, the status bar
icons are shown, only be to be hidden shortly after the
animation ends when we receive a callback from WM that the
top app hides the status bar.

This was an unexpectedly difficult fix as there's a brief
period between the launch animation ending and System UI
receiving a call that the top app wants to hide the status
bar, where there is no existing state that can tell us that
we should be hiding the icons. A similar case is currently
handled in StatusBarHideIconsForBouncerManager by posting
a callback delayed to not hide/show icons within 500ms, but
that's not particularly robust.

Fixes: 257292822
Test: manually launch camera
Test: atest CollapsedStatusBarFragmentTest
Change-Id: Ib54917b4e65529a2fd809eb67159d57d64e8097f

Former-commit-id: 0abf06dcfb70f34f12fb3ca3d0dc432a770a5e7a