Enable network type icon overlays by carrier Id
This CL builds out a new configuration resource
`carrierid_icon_overrides`, and a format that allows for overriding a
known network type with an arbitrary icon.

We store a mapping of all known carrierId overrides that exist, and
check for an override for the `carrierId & networkType` combination at
runtime.

This CL also caches the calculation in SystemUI since it represents a
new expensive calculation potential at the time of display for the icon.

Lastly, add demo mode command parsing for carrierId

Test: atest MobileStateTest; atest NetworkControllerDataTest
Test: manual using `-e carrierid <id>` network subcommand
Bug: 241996752
Change-Id: Id0e8acf179db5c3a70f4caf31e58f8d8040690a0

Former-commit-id: bb7f96387d0a035c2f0e7ac19675012b854f6623