[User Switcher] Fix username sync in bouncer.
The user name is not updated when userinfo changes (name and avatar).
Updated usertracker to listen to this intent and updated repository to
listen to these changes. Corresponding tests have also been updated.

Fixes: 254678959
Test: Observed username in bouncer. I unlock the device and change the
username for the current user. Turn device screen off and on and go back
to bouncer. Observe that the username is the same as the one changed in
settings.

Change-Id: Ie6e289dfc65c14f9b0a35f031c46122d10dca9d0

Former-commit-id: d1638a8b626aa861d3b11821becf39e0eb2db0e5