[Media TTT] Handle multiple devices
Handles multiple devices so the view of the most active recent device is
displayed. Adding a queue to store the active device states. Whenever
there is a new device is trying to transfer media, it has a certain
state in the queue and stayed in the queue until it gets removed when
remove view is called.

Bug: 213576772
Test: atest ChipbarCoordinatorTest
Test: atest TemporaryViewLoggerTest
Test: atest TemporaryViewDisplayControllerTest
Test: atest MediaTttChipControllerReceiverTest
Change-Id: I3b2502f045281d01b51215d58bbb7d3759d04de2

Former-commit-id: b69b45c6cf9ea468056d4f95a3d88c5fa6b8d6b3