Fix an issue where the highlight for drag to split was incorrect
1. Have a fullscreen immersive app open
2. Swipe bottom of the screen to show the taskbar
3. Drag and app

In 3 when you start dragging the app, the taskbar hides because you're
in an immersive app, this causes the rect for the drag targets to
be different, so we'd end up animating the drop zones to switch even
though the type of the target was the same. Fix is to check that the
target type is different before animating the switch.

Bug: 255008704
Test: manual
Change-Id: Ic8a7b23e79c54ba74aa6bebf528fb0b3894a8fa1

Former-commit-id: d43f418ffdd4e43e42224b5578585cae98b8762b