Remove AndroidPackage#isSystem
Follow up to previous change I2a915e55f609992729e082815d78fd2d73cb79fd
moving the partition methods from AndroidPackage into PackageState.

This moves the remaining usages of isSystem so that all of the
partition methods have now been migrated.

A few usages of AndroidPackageUtils methods exist for cases where
an AndroidPackageHidden cast is necessary to read the field
directly from the AndroidPackage object. This is usually because of
the install flow, which doesn't guarantee a PackageSetting until
late in the process.

API-Coverage-Bug: 262593427

Test: presubmit

Change-Id: Iff6b800de68fee92e29eb7a0410a5232386f8c19

Former-commit-id: 7c3d904fc278c556545683ed3e33111b1b36f9bb