Change volume ringer icon based on device config
When device config volume_separate_notification is enabled, change the ringer drawer icon, so it covers both ring and notification, and not just one. When not enabled, the ringer icon should remain/switch to the old bell icon.

Also, the icon on the secondary vertical progressbar follows the device config flag, matching the horizontal slider for ring:
https://screenshot.googleplex.com/6N5CPjSCR4P5DbZ.png

Bug: b/38477228

Test: atest VolumeDialogImplTest

Change-Id: I2d1d7348d3d6506fda98394a2ddf18461f575fc2

Former-commit-id: 903b3025a7e5caaddae848376bd12b62e698315c