Fixing scrims visible in split shade on lockscreen
This issue occured when we were going to displayed lockscreen directly from expanded shade - compared to turning off the screen and then going to lockscreen after a moment.
Animation didn't have time to fully run and we ended up in not fully clean state of QS. This change forces closing QS and reseting its state when going from unlocked mode to keyguard.

To make it easier to catch similar issues in the future I also added wtf logging which is triggered when we recognize scrims and lockscreen clock visible at the same time.

Also removing mInSplitShade which was just duplicate mSplitShadeEnabled.

Test: NotificationPanelViewControllerTest
Test: display heads-up notification -> long press power button -> lockdown -> see scrims not visible on lockscreen (see b/257930534 for more reproduction steps)
Fixes: 257930534
Change-Id: I79fc78c24d4755d3fc2bbdcf5f4438993a72706b

Former-commit-id: b8ec0378df1a55b06c4aaef5bccfa0ad982127f8