[Bouncer] Update bouncer view with display size.
Updates the bouncer view when display size has changed and when font
scale has changed. When this specific configuration changes, we
invalidate the cache of the view flipper and tell each view mode that
the configuration change has been made. In the UserSwitcher view mode,
we reinflate the user switcher.

Fixes: 257513443
Test: Change display size, turn screen off and on and observe security
view and user switcher has changed size. (On Pin, Password, and pattern)
Test: Change font size and bold font, turn screen off and on and observe security
view and user switcher has changed size. (On Pin, Password, and pattern)
Test: Added unit test.

Change-Id: Iffbbe6f5b0f32f4b40dd70e9bd534b444a8d1ef6

Former-commit-id: fb36c48bbd204b77ee4a4528148698f1b973f1a4