[Media TTT] Displayed View Keeps Screen On
This CL ensures that the screen is kept on when the chipbar is
displayed. It acquires a wakelock with different levels and flags when
the screen is already on. This doesn't affect the screensaver mode.

Bug: 241929425.
Test: Ran adb commands when the screen is already on, and made sure that
the screen goes off after the view is removed. This was tested on
a phone and tablet devices.
Test: atest MediaTttChipControllerReceiver.
Test: atest TemporaryViewDisplayControllerTest.
Test: atest MediaTttSenderCoordinatorTest.

Change-Id: I2c471c435d297c074ee792c06c80112de24ac5f2

Former-commit-id: 203a6745aa044efd5952b2baaf63441badb98522