Remove isInstrumenting workaround from receiver flag enforcement
Android T added support to register a runtime receiver with a flag
indicating the receiver should not be accessible to other apps on
the device. To ensure developers could take advantage of this new
flag, receivers registering for non-system broadcasts and targeting
T+ were required to specify a flag indicating the exported state of
the receiver. While flags were being updated in the platform, an
isInstrumenting check was added that automatically flagged receivers
as exported if they were under instrumentation. This was intended
to allow additional time for these receivers to be flagged, but
this workaround caused issues to be missed with WebView and
potentially other components. Enforcement of this flag was punted
from T and is now in plan for U, but before enforcement is enabled,
this commit will remove the isInstrumenting workaround to ensure
tests run with the same behavior as the platform and any potential
issues can be caught earlier in the cycle.

Bug: 234659204
Test: atest CtsWebkitTestCases
Change-Id: Id24e15f951901b4d998f8301950d49bad53ca278

Former-commit-id: 297676117569ba86b4c829724ae350db41e4e479