[pm] Implement the BIC service
The BIC service monitors two categories of events:

1. usage events to tell the foreground/background state
of package installers.
2. packages install/uninstall.

Based on the timestamps of the above events, the BIC service
detects the background installed packages.
The BIC service also stores the list of background installed
packages on the disk.

The clients of the BIC service can query the list of background
installed packages.

Bug: 238451991
Test: BackgroundInstallControlServiceTest
Change-Id: I58c98a176897893b60cf24d01c69266771d102ca

Former-commit-id: 26e8370b96abe441eae6c0ec5700cff1eddcebb9