Support back gesture to split pair in background
Split screen will be dismissed right after both sides of the split
were occluded. So if users try to use a back-gesture to go back to
the previous split pair, it’ll be back to home instead of going
back to the split pair. This CL support the split pair in background
to make split experience better.

Bug: 236317871
Test: atest com.android.wm.shell.SplitLayoutTests
Change-Id: I1ead65011ced82e1faa24ccfbc8bf474828fc7dc

Former-commit-id: 76ce5869fc9acbd8b925682d99dc73833bc5aa2c