Fixing scrims jumping when media player state changes
The issue was caused by mKeyguardStatusAreaClipBounds not being emptied when no longer needed.

In portrait when media is appearing or disappearing in QQS, we animate notification scrim to its new position. This animation was dependent on mKeyguardStatusAreaClipBounds (now renamed to mLastQsClipBounds) that was changed only in folded mode. In unfolded animation was using old values from that field which were never updated. Solution is to always keep last clips bounds updated and use it when animating media visibility also on large screen portrait.

Fixes: 234129271
Test: Have media running -> fold then unfold back -> open shade in portrait and remove media player from QQS -> notification scrim should animate nicely
Change-Id: I930456ac79095eda55e098e02abf96b45dec1214

Former-commit-id: eb21e52b88c5184d9d777534b8024804525f76d2