Make ImageWallpaper handle WallpaperService messages on its own thread
When a wallpaper is called to be attached, IWallpaperEngineWrapper
sends a DO_ATTACH message to be handled on the main looper of the
wallpaper implementation. In the case of ImageWallpaper, it is
currently systemui. When system ui is occupied by other tasks,
reception of this message is delayed. And this decreases the
performance of switching between Android users in case the target user
has an ImageWallpaper.

This CL makes sure these service messages are handled on
ImageWallpaper's own thread, and the drawing of the wallpaper is not
delayed.

Bug: 242969351
Test: Manual comparison of perfetto traces
Change-Id: Ia911f1b879a20658d0b71e3486c079cbbcb081c8

Former-commit-id: 696adad6149192bd0e322c8a21fba4010cc4ce29