[Bouncer] Set message for dismiss api.
Message being set to error message from viewmediator callback and prompt
reason from view mediator callback are overridden from the setText that
is called in #appear.

Only set message when text is empty. Additionally, for the modern
bouncer, the prompt reason was not updating properly. This is because
the old bouncer has a prepare method that udpates the prompt reason.

Test: patched in http://ag/20257883 and did dock setup. Observe with
MODERN_BOUNCER flag on and off. Both update the message.
Test: Add fingerprint auth and restart device to activate
REQUIRES_HARD_ATUH. Observe with
MODERN_BOUNCER flag on and off. Both prompt to unlock the device before
using auth.
Fixes 254841665

Change-Id: I8db831bfec4f6ac28f07049669469fdbb896de06

Former-commit-id: 2e482ce2349379e1d2160a8dae0f26256ecebb37