Override registerReceiver methods in test Context to accept flags
The BroadcastInterceptingContext allows a broadcast Intent to be
intercepted in the local process and sent directly to the receiver's
onReceive method. This commit updates this class to also override
the registerReceiver methods that accept the flags parameter to
ensure broadcasts can be sent to receivers registering with flags
during tests.

Bug: 234659204
Test: atest com.android.networkstack.tethering.EntitlementManagerTest
Change-Id: I9366ed1a3eb5f370b93e72981abac535f675a2bb

Former-commit-id: f73b5584e2f7c0fd68d95dd47cc4bf6415d52773