Fix prefill request is blocked waiting for an inline request.
When asking the IME for an inline request, the IME will check if the
autofill id of the target view matches the id of the current IME served
view, otherwise the IME does not respond inline request to the Autofill.
When an activity starts, the id of the IME served view is reset to 0.
And the target view, Autofill used to do the pre-fill request is the
root view, sometimes its id is not 0. Then the request is blocked to
wait for an inline request because autofill does not have a timeout
mechanism.
To resolve this issue, use AutofillId.NO_AUTOFILL_ID to make sure the
id of the pre-fill request is 0.

Also this change remove the duplicate code about the
notifyViewEnteredLocked for normal view and virtual view.

Bug: 258532947
Bug: 253180120
Test Manual
Test: atest CtsAutoFillServiceTestCases

Change-Id: I6962ab6d844fd8b7c9ba778ac40630586ef73c39

Former-commit-id: 78bdebcf6a24bb133532987c705709f745823000