Reduce Accessibility animation-wait timeout
Logic on the client-side often uses a similar mechanism
(waiting for no-changes) to detect idle-ness. The most widely
used (esp. in tests) is UiAutomator which happens to share
the same 500ms timeout. This means there is a race between
the timers+binder where, accessibility will have up-to-date
information, but not send it because it is still getting
updates from surfaceflinger. To prevent data from getting
too stale, lower the timeout here so that the client can
receive the latest window information before timing-out and
using stale information to perform other operations

Bug: 258073789
Test: atest NexusLauncherTests:AddConfigWidgetTest#testWidgetConfig
Change-Id: I7c32acbc36810d74fe1baff538e26f0899a730aa

Former-commit-id: f404193c915456e1ea815461689991575179e460