Bouncer - Remove cases that forced it to stay open
Until recently, when the user swiped up from the shade, the bouncer
would display. This required extra care when listening to panel
expansion events, to make sure the bouncer state was correct while
swiping. Now that the bouncer does not display in this manner, we can
safely remove a few checks regarding bouncer state that were
preventing it from hiding in some scenarios.

Fixes: 258588837
Test: atest StatusBarKeyguardViewManagerTest
Change-Id: I4fb6c38de3d309299b01d213ebdcf00b1549f1d9

Former-commit-id: 00bd9b3187ef6d6721234418b79f393e96131e87