Distinguish between selected panel or structure
To pave the way, we add a SelectedItem that could contain a
StructureInfo or information for a Panel.

In particular, a Panel points to the ComponentName of the
ControlsProviderService (and not of the activity) for backwards
compatibility and migration. As we still need to match the SelectionInfo
that has the activity information, this is not an issue.

Also, add persistence for the panel using the same keys (and an extra
boolean) for better migration.

Test: atest com.android.systemui.controls
Test: atest DeviceControlsTile
Fixes: 258442632
Change-Id: I544f5cdcb4fdfe0488df05d686173ed3bb709b8e

Former-commit-id: 82587b1a426f84544568303d67f859f2aaad6037