Fix 'Generic.kl' file has no 'KEYCODE_FOCUS' mapping
This keyCode is defined in both 'keyEvent.java' and 'keyCodes.h', but not in the Generic.kl file. The comment of 'KEYCODE_FOCUS' says that this keyCode is used for camera focus, so this keyCode should be sent to the camera app, but now he is not in 'Generic.kl'.

Change-Id: I2d0a49edae0390c9199efe0f01e527b15d14e68e
Signed-off-by: zhangzhihao7 <zhangzhihao7@xiaomi.corp-partner.google.com>

Former-commit-id: e95a4454a366ba62dafee0d4a4084af352b5fd91