Call a search activity when using KEYCODE_SEARCH
This change adds the ability to change the behavior
of the search key to open a search activity instead
of attempting to start the search activity of the
currently opened app. The new behavior remains the
same by default, but now it can be changed to
either do nothing or open a configured activity; in
this mode, if no activity has been configured,
the `ACTION_WEB_SEARCH` Intent will be launched.

Test: manual
Bug: 240865724
Signed-off-by: LiZhihong <lizh52@lenovo.corp-partner.google.com> DingYong <dingyong2@lenovo.corp-partner.google.com>
Change-Id: I3037d255b693ade3f45e170746e0c2b2de5db6fb

Former-commit-id: c978736155c945e5b173dadee5ad4ad3e2b6fb40