Check home controls availability on overlay active.
The callback that home controls observes is only triggered when the
controls services update, which is one of the conditions for whether
home controls is available. When the user adds the first or removes the
last device control, the hasFavorites condition changes but does not
trigger an update.

So to address that, home controls should check availability when the
dream overlay becomes active, parallel to which on the lock screen.

Test: DreamHomeControlsComplicationTest
Test: reboot device with no favorited device controls > add favorite
device controls > start dreaming > see home controls chip
Test: similar to above but start with favorites and remove last
Bug: 257348199
Fix: 257348199

Change-Id: I537a240963e0523c8fc40008dee961c7a434094a

Former-commit-id: 4c5add5c98fe7c3e33cd3bfe0160f1c38041010f