Incremental update to portrait split gesture animation
* More closely trying to match PreviewPositionHelper to
the calculations in PortraitPagedViewHandler.
* It's correct for unstashed, it's improved for stashed, but
there's something still slightly wrong.
One thought is to match the calculations line for line,
similat to setSplitTaskSwipeRect(), using the screenHeightPx
instead of canvasHeight to determine the scale (doesn't work by itself),
but it's slightly closer, though we'd also need to make changes to
measureGroupedTaskViewThumbnailBounds() accordingly

Bug: 219411750
Test: Visually inspected
Change-Id: I518a884552accfd1c56c14cf98fd0fdcd00e817a

Former-commit-id: c702c4f542a3e7d75fdab7af53578efbb8fc0114