Add ability to AND and OR conditions together
In the dock setup flow, we have complex triggering which could be made
simpler with the ability to OR conditions together. To cover the case
where the value is undefined, this change introduces 3-valued AND/OR
operators which can be used to construct new conditions from an existing
condition.

To keep the syntax readable, we add or() and and() methods to
conditions, so the caller can simply say condition1.or(condition2)

Bug: 255372044
Test: atest com.android.systemui.util.condition.ConditionTest
Test: atest com.android.systemui.util.condition.ConditionMonitorTest
Change-Id: Iac583492669cbb573c84807c9c0e78c35dfc02a0

Former-commit-id: fb2123974d5146b8f19f2f53591925d0de5136db