Split InternalInsetsType into InsetsType and the ID of InsetsSource
The is a step to remove InternalInsetsType which is used to categorize
insets sources and is also used to identify the corresponding objects
like InsetsSourceControl, InsetsSourceConsumer, ..., etc.

This CL migrates the usage of categorizing to the public types, and uses
the internal type as the identifier for now. Eventually, the ID will be
specified by the source producer, and we won't need the pre-defined ID
then.

Bug: 234093736
Test: atest InsetsSourceConsumerTest InsetsSourceTest
      DisplayImeControllerTest CompatUIControllerTest
      CompatUIWindowManagerTest ActivityRecordTests DisplayPolicyTests
      ImeInsetsSourceProviderTest SizeCompatTests
      WindowContainerInsetsSourceProviderTest WindowContainerTests
      WindowOrganizerTests WindowStateTests
Change-Id: Id574f98af18ab95b11fed13a09e08bc6a7929957

Former-commit-id: 71713dd50e0691df66e180b334b4b6d23d12abf5