Tell input System Application Overlays are trusted
They can only be created by applications holding
SYSTEM_APPLICATION_OVERLAY permission, which is signature|recents|role
and it was specifically designed to only be used by compontents tightly
integrated with the system, such as device intelligence and digital
wellbeing.

Without this change SAO overlays are in a weird state where they can
stay visible on top of apps that are hiding overlays, but the are
generating FLAG_WINDOW_IS_PARTIALLY_OBSCURED touch events, causing
security sensitve apps to drop the touches.

Bug: 199325240
Bug: 200938390
Test: Manully verified that system caption doesn't block interaction
with permission dialog
Test: atest CtsWindowManagerDeviceTestCases:HideOverlayWindowsTest

Change-Id: I2c88bcd1d064758b901dda7d0ef30bc92b9cd56d

Former-commit-id: e8be9e6c628438be499d00ec6d9194ce69c03086