Add an enum for decisions about whether to launch FSI.
This change splits shouldLaunchFullScreenIntentWhenAdded into two parts: one that returns the specific decision for whether to launch a full screen intent for an entry, and one that logs that decision.

It also adds a sub-category of "suppressed by DND" called "suppressed only by DND", where DND is the only thing stopping an FSI from launching but all other cases would allow it.

Bug: 248325248
Test: NotificationInterruptStateProviderImplTest

Change-Id: Idd1299312edb58d005e089d89db7921c15c6e2bc

Former-commit-id: 7c1e219d6d5ef54cf9377c1b60c4fee1ccae0fe6