[Bouncer] Remove flicker of message area.
When locked out for 30s, we add a timer that changes the message every
second. The animation is a bit jarring, so I added a way to disable the
animation in cases like this.

Fixes: 257407393
Test: Add unique incorrect pins until I get the dialog that I have been
locked out. Observe the changes. Also observe that the bouner title
still animates in other cases.
Test: Added unit tests.

Change-Id: I5e09686174ac17cae139f5f9535e946e39d6d009

Former-commit-id: 0b9bc54cb304472a19436ed106bb866859c97bcc