Factory Reset Protection via MainUser
Instead of using the SYSTEM user for Factory Reset Protection, use the
Main User (if it exists) instead.

This means that, even in Headless System User Mode (HSUM), the MainUser can set the FRP.

Part of this involves migrating SECURE_FRP_MODE from a per-user setting
to a Global setting. It should always have been a Global setting, since
it is a device-wide property. It should only ever have been set or read
on user 0 prior to this cl anyway.

Bug: 249097735
Bug: 256624031
Bug: 159629501
Test: manual: In HSUM, user 10 has account+PIN; factory reset; SUW demands credentials post reboot
Test: atest FrameworksServicesTests:SyntheticPasswordTests FrameworksServicesTests:WeaverBasedSyntheticPasswordTests
Test: atest CtsPackageInstallTestCases:android.packageinstaller.install.cts
Change-Id: Ic84cb9c3b74317813d7ae11cdae2eba729d8fa88

Former-commit-id: caa07ff5286a79dfb7492201f601b56b148e7dcd