GraphicsEnvironment: Determine storage limit for EGL blob cache
The application has a quota provided by the platform informing it
how much space it can use for temp storage.  The value can change
over time based platform resources.

Use this value to dictate how much space drivers can use to store
persistent entries in EGL blob cache.

For games, allow 50% of the storage to be used.
For non-games, limit to 25%.

This CL works in conjunction with a native change to use multiple
files for EGL blob cache. (ag/20505145)  The limits for monolithic
cache are unchanged.

Test: ANGLE trace tests, various games and apps
Bug: b/246966894
Change-Id: I38ee45a27c29f913f5de9c0d76722da0d1dcade3

Former-commit-id: 74737915b1404de72e3dd208ac03420612adec08