[Unfold animation] Get animation scale from content provider
ValueAnimator#areAnimationsEnabled() method does not
reflect the latest state of the animation scale
at the time when content provider emitted an update.
It led to a bug after disabling and enabling
animations: we added the unfold overlay in the
 UnfoldLightRevealOverlayAnimation (as animations
enabled) but didn't remove it because the transition
progress provider never reported animation start,
progress and finish events.

Changed to use the content provider to get
the latest state of the animation scale.

Bug: 256072006
Test: manual fold/unfolds after enabling/disabling talkback
Test: com.android.systemui.unfold.util.ScaleAwareUnfoldProgressProviderTest
Change-Id: If5d982791b9f8103ecfd8db44e24d2486560d4f5

Former-commit-id: 982c56f6613e5a5058550141296f7a8310dbf960