[Unfold animation] Move the dark scrim below
How rotation animation works:
* It captures a screenshot of the screen
* It puts the screenshot over everything on the
  screen (in a container with z-index
  Integer.MAX_VALUE)
* Starts the process of resizing windows to the
  new size. Because the screenshot on top of
  everything, the resizing is not visible.

Black unfold overlay is also added with
z-index Integer.MAX_VALUE but because it is
added later it is rendered on top of the
rotation screenshot. So it is visible how
the overlay is resized.

Updated z-index of the unfold overlay to render
below the rotation animation screenshot to hide
it while we are rotating.

Bug: 258213952
Bug: 255955520
Test: manual fold and going to tabletop mode after rotating
  (see comparison videos in the CL)
Change-Id: I5e9b5c72dfa28e4ae9f447ad8df7b403c97bf43e

Former-commit-id: 0ea7ac03023cc042763711989dc8b666eccdfeaf