Unhide clearCredentialState framework api.
This API is independent of the get-credential api because at
the time of the get call, information about whether the previous /
existing user state should be cleared has already been lost. That is,
we can't have this as a parameter to the get call because it'd be
impossible to set it at that point.

Bug: 246564035
CTS-Coverage-Bug: 246637346
Test: Local Build & Deployment

Change-Id: Icc12a89039e7088bcee8fd6ae20a6349f730ac91

Former-commit-id: 1db4fabbebb9da32e420b5fd1a9d09f3b1868277