Fix bug of keyboard hidden after rotating to landscape mode.
Add logic in IMS#onShowInputRequested: if IMS receives a Implicit show request and the keyboard is already shown in full screen mode, there is no need to hide it.

Otherwise it would hide the keyboard after rotating to landscape mode, because the rotation causes recreation of the activity and triggers auto-show behavior of Ime.

Bug: 256739702

Test: manually test

Change-Id: I195c2c6bd6f6815806dc90f0a386b7c0bccee7a1

Former-commit-id: 9e3642a928042aef03c2686de485bfcd5bb34af0