Clear underlying network info in onDefaultNetworkLost()
Fix the wrong behavior that mUnderlyingNetworkCapabilities and
mUnderlyingLinkProperties is always cleared when
handleSessionLost() is called.

mUnderlyingNetworkCapabilities and mUnderlyingLinkProperties are
the info of underlying network, they should be cleared only when
the underlying network is lost. But the current design will
always clear them when handleSessionLost() is called, which will
make the VPN app won't get NetworkCapabilities and LinkProperties
of underlying network when VpnManager event is sent.

Bug: 237050331
Test: atest FrameworksNetTests:VpnTest
Change-Id: Ibfd4f0199eafb67bfddb4ee0664550d3866e1840

Former-commit-id: 4eac85accf919f749d852e13134b96564051966c