Respect "nosensor" and "locked" with ignoreOrientationRequest being set
In this implementation, we always resolve the orientation for a
component as the ignoreOrientationRequest was disabled. We then
use the actual flag when deciding the display orientation and when
querying from a descendent to handle the orientation change itself.

Main use case why this is important is Camera apps that rely on those
properties to ensure that they will be able to determine Camera preview
orientation correctly

Fixes:254691921
Test: run `atest WmTests:ActivityRecordTests WmTests:TaskTests
      WmTests:WindowContainerTests`

Change-Id: I96dca18a561e2acf961a462d3a4db484f003be6f
Merged-In: I96dca18a561e2acf961a462d3a4db484f003be6f

Former-commit-id: cc2d539e801ab7aa748c780c41c1a1569bdd112c