Ensure final height of UMO transition is the hostView's height
Since the targetHost should be measured correctly, and it sets
the child view's size when it is added (either by remeasuring
or by directly setting it), setting it again to the target bounds
is unnessecary and potentially buggy.

Additionally, when we obtain a view state, we always obtain it
at an expansion of 1f. This prevents any squishiness from being
applied when remeasuring.

Test: Manual
Bug: 232528163
Change-Id: I0a4e6c457663dfe376ad7754f6d151f47d3aeaf3

Former-commit-id: 876efc4b26152f3f48d6a195ceafc5bf4e36a34f