Use VirtualInputDeviceConfig subclasses to create virtual input devices.
These config classes encapsulates the config parameters. This is to keep
the argument list of createVirtualFoo APIs from growing longer every
time we add new config parameters to an input device. Only shared common
configurations such as product ID are held in VirtualInputDeviceConfig.
Optional parameters such the sizes for touchscreen will be held by the
corresponding subtypes dedicated for that input device. This way we keep
what is settable for each device clear.

Bug: 237537306
Test: atest CtsVirtualDevicesTestCases CtsHardwareTestCases

Change-Id: I609d109891fe871367869c730ed1905be8889f89

Former-commit-id: 290b4fd61abb6040004d7c19df4ee1781d17d4c4