Remove the usage of givenContentInsets for non-IME windows
Currently, no insets is calculated based on contentInsets and other
fields set in WindowState when the InsetsFrameProvider is given. It is
safe to remove the calculation for windows other than IME.

The fields in WindowState, including touchableAreaInsets, visibleInsets,
contentInsets is a description of the window itself, instead of the
insets it providing to other windows. Such that, we can remove the
consideration of those fields for most cases.

Bug: 202360061
Test: DisplayPolicyInsetsTests
Change-Id: Ie127f4655ed923545dede16ebc615dbf8384d34a

Former-commit-id: 77a0bd889978e4b9945f146ce7fac212ff0540ef