Remove usage of GlEngine
The new Engine of ImageWallpaper, a CanvasEngine, was enabled in droidFood under a flag. It has been tested for several weeks and is ready to be enabled by default, wihtout any flag.

The ImageWallpaperTest has been slightly refactored to fix all warnings and reintroduce two GL tests that were ignored.

A very small fix was added in ImageWallpaper.java: if the bitmap fails to load (which should not happen) for another user than USER_SYSTEM, it will be reset to default for this user and not for USER_SYSTEM.

Test: atest ImageWallpaperTest
Test: atest WallpaperLocalColorExtractorTest
Fixes: 243768810
Bug: 243402530
Bug: 254512923
Change-Id: I0bfbe9b6aee7a4da349b412399196504511b69f0
Merged-In: I0bfbe9b6aee7a4da349b412399196504511b69f0

Former-commit-id: 38545343c6ed615c56f0570e80bfd767abf824de