Move surface effects package to AnimationLib.
in order to share surface effects between launcher and sysui, the
package has been moved to the AnimationLib, which is visible to launcher
as well.

fyi the corresponding tests remain in the SystemUiLib module.
technically they should be moved to AnimationLib together, but those
tests are dependant on SystemUI testing frameworks (e.g. SysuiTestCase,
FakeExecutor, etc.). because of this, some of the variables are marked
as VisibleForTesting.

Bug: 249536060
Test: MultiRippleControllerTest MultiRippleViewTest RippleAnimationTest RippleViewTest TurbulenceNoiseControllerTest TurbulenceNoiseViewTest
Change-Id: I37cdc7cf9064d4d8f02afe72ad16b1cc3df95b6c

Former-commit-id: 7b8e5271340bc67bdee84583f7d3350fabfac01b