Fix clock being too low in AOD
To fix the clock being too low, this undoes the relevant portions of
I3feea196bf348630697385e010a13955cdb0a897. We'll want to revert this
when we figure out why the clock is moving so low in AOD.

The clock still moves vertically a significant distance when entering and leaving AOD, but not so much as to overlap with the UDFPS/lock icon. I'm continuing to investigate what's causing that.

Test: Manual
Bug: 259758161
Change-Id: Iaeddce2d99a2b8db5afb6ac8ca508eeab3ae223e

Former-commit-id: 4714ab131716246a4e94c62c75e63df4e7f3eb90