[Bouncer] hide/show keyboard when focus is...
regained. For the bonuncer, often times the UDFPS overlay, or AOD can
take the window focus. This will prevent the window of the bouncer from
sending signals to the input manager to hide/show the keyboard. Add an
override method for the password view to show or hide the keyboard
contingent upon the visibility of the view. Also refocus edit text when
changing orientation of device.

Fixes: 259552291
Test: Open scrimmed bouncer from AOD.
Test: Open bouncer and hide bouncer a bunch.
Test: Open bouncer from udfps overlay.
Test: Unit test.
Test: change view mode from landscape to portrait when on bouncer on tablet.

Change-Id: Ic30099408086f6dcc63506cac6ca2fb015504d17

Former-commit-id: 9ce60a5ab86d443ebe71b3e8681df7de004d729b