BinaryTransparencyService: Log new atom to Westworld.
Per the introduction of a new atom (MobileBundledAppInfoGathered),
we are logging new pertinent data to Westworld. These includes:
- preloaded apps that are updated
- new MBAs that are installed in the background

Note that we are *not* recording the full set of preloaded apps
on a device due to being cautious about data footprint. Instead,
we rely on those being implicitly covered under the vbmeta digest
measurement.

Bug: 245944158

Test: Manual. Set up the device by going through the SuW. Connect
to the Internet and allow for updates and also let the system
install several MBAs. Then, invoke `statsd_testdrive 566` from a
terminal (as 566 is the atom number). Within 60 seconds of that
invocation, trigger the package measuring job via the following
command from another terminal:
`adb shell cmd jobscheduler run android <job-id>` where the
job-id can be found in logcat.

Change-Id: I148dc6a7f1c8362e87321ba5d4879b266f1f11c9

Former-commit-id: 0000111139a3104d75a7ca977ffae5fb42942871