Inject back key event if focused window didn't be drawn
If an user applied a back gesture and core found the focused window
without any surface drawn, the input window could still be invisible
and the focus request would be pending. In such situation, we should
directly inject the back key so it could dispatch to the right focused
window after the focus request is applied.

Test: atest BackNavigationControllerTests
Bug: 260265858
Change-Id: I77b1cad23d458927715333bdcef701f5cf409347

Former-commit-id: a13a07e395922ec215bd4d0f3ba301d605009ecb