Use parent frame to update wallpaper offset
This partial reverts [1], which is similar to restore to [2].
Because shell rotation transition was disabled and currently
shell also doesn't control wallpaper, the corresponding changes
can be restored to avoid side effects. Such as after changing
screen resolution, the last reported bounds may be still old
size, which leads to a temporary additional offset.

If shell rotation transition may be re-enabled in the future,
the correct fix should be to stop updating window frames of
invisible requested window instead of trying to use previous
bounds to calculate, i.e. keep parent frame in previous rotation.

[1]: I1db9207c79bc45dedd90e7b13af5517f09abe4a7
[2]: Icf55e3570c260b963fab6dc85814f9e8e305163c

Bug: 264018192
Test: WallpaperControllerTests

Change-Id: Ia6bf26b4576dcf9a036c3b8686e015e19fbf57a2

Former-commit-id: dbfa31c2b48c94cfc03219f96e171899fc8cc773