Allow non-updatable modules to contribute to @TestApi
java_sdk_libraries can be part of non-updatable modules. Those modules
might have a need to expose @TestApis that are only used in CTS tests.

This change adds an ability for them to do so by introducing
all-non-updatable-modules-test-stubs, and adding it to the
android_test_stubs_current.

Note that this change doesn't allow updatable modules to contribute to
test API stubs, since the list of the modules that can contribute to
test APIs stubs is hardcoded in the non_updatable_modules constant.

Bug: 261004711
Test: m
Change-Id: I9d17f49702bd64f5718b445f14b203c146bc6794

Former-commit-id: 5593fbbca8b5d2dd3a4b76a0a5623c88b577c908