Support new pipeline demo mode in the old view presenter
Adds support for the new mobile pipeline's demo mode to the old front
end. Note that this is kind of hacky because we actually don't have to
do anything to support demo mode, since it's implemented entirely at the
level of the repository. However, due to the fact that the old
implementation of demo mode added a special view, `DemoStatusIcons`,
that overlays the regular icons, we have to suport the new pipeline in
_that_ icon container.

Test: manual
Bug: 249790009
Change-Id: Ib02a1eb6f6abaec113a15045f9bd0b151d2a773f

Former-commit-id: cc43d6a8782210bd2284114e7e805078958749c4