Support shared isolated processes.
This introduces a new flag, Context.BIND_SHARED_ISOLATED_PROCESS,
which allows a single client to bind multiple isolated services into a
single isolated process. This allows the system to save resources when
multiple isolated services can safely be co-located into a single
isolated process.

We reuse the existing instanceName parameter to specify the name of the
isolated process. After the first bind for a client with a particular
instanceName, subsequent service binds with the same flag and
instanceName will result in the service being bound into the already
existing isolated process.

These are scoped to the client app; it is not possible for an app
to bind an isolated service into a shared isolated process that
belongs to another app.

Bug: 243554393
Test: atest ActiveServicesTest
Test: atest android.externalservice.cts.SharedIsolatedServiceTest
Change-Id: I6b88c551082be849f384849955d3a43e0aa707c1

Former-commit-id: 72d234ea09b428032d3e89cf2de101f75352c705