Fix errant bouncer/keyboard display over keyguard
Legacy logic attempts to infer whether the bouncer should show based
upon panel expansion events in a variety of scenarios.

Modify the logic a bit to require show() be called on the bouncer, in
order for any events to be forwarded.

Fixes: 260589442
Fixes: 260310219
Test: atest StatusBarKeyguardViewManager
Test: manual - Bouncer works for SIM, from AOD, from UDFPS, over
occluded activity, while unlocked, over dreams, with face auth,
tapping on notif

Change-Id: I4e5919d222c1c94731dc5bf050ff66a8376b38a9

Former-commit-id: d49175bc73b2b076431066498077928a77524f02