Reduce operations in postApplyAnimation for shell transition
This is a partial alternative solution of [1] to avoid black background
(no wallpaper target) when turning on screen with a translucent activity
which is occluding keyguard while AOD is disabled.

[1]: Ifc3ce28424fe87b1e0a0a2af07bdad559575ebec

Now instead of always recomputing wallpaper target after committing
visibility, the update hint only needs to be set when there is a
translucent occluding activity.

Also use usingShellTransitions to skip some legacy animation checks
that shell transition doesn't care.

Bug: 260059642
Test: Enable shell transition. Disable AOD
      KeyguardTests#testDialogShowWhenLockedActivity
      KeyguardTests#testTranslucentShowWhenLockedActivity
Test: Transition latency of OpenAppMicrobenchmark
Change-Id: Ic3c704f0579690c51f7188a26c5256dc3e74e1f7

Former-commit-id: 607c7a191434b58e44b4d727dd6a5ba936653d1e