Move Lifecycle out of DreamOverlayModule.
Lifecycle is used by both dreams and complications to understand their
current state. The lifecycles of these two components can be the same,
such as when the complications are part of the dream overlay. However,
it is possible the lifecycles can be unrelated in the future, such as
complications being hosted in another SystemUI surface.

This changelist addresses this possibility by moving the current
lifecycle out of the DreamOverlayModule and instead injecting it into
the DreamOverlayService. In the case the service creates the
complication layer, this lifecycle will be bound to the appropriate
component. This change also introduces a concrete implementation of
LifecycleOwner, which exposes the registry.

Test: atest DreamOverlayServiceTest
Bug: 261781069
Change-Id: Ic04f62a89405b835c397b1ce348780d8aa2f05c1

Former-commit-id: 6eb501e6c6766c71b7446651f58964ee9042474d