Fix ActivityEmbedding animation with new split in different layout
When launching an activity into a new TaskFragment with a different
split layout from the existing split, there can be many situations, such
as different split ratio or layout direction.

Use a jump cut for now since it is not a common use case, and it is hard
to cover all situations.

Bug: 241043533
Fix: 258126611
Test: verify with opening new activity to new split layout.
Change-Id: I6c96518b282567339fb890e7f4af7d9d82c98418

Former-commit-id: f2296dcb79ddd7225172aad4c4809f54ca3faee7