Fix parental control dialog
This CL fixes a bug introduced by ag/19712818: when inflating the
parental control dialog, we should use the Quick Settings context (which
is always using dark mode) and not the application context. Otherwise,
if we are in light mode, then the color of the text is the same as the
background and it looks like there is not content in the dialog. In dark
mode, everything was looking fine because the text color from the
application context would be light.

Bug: 260108168
Test: Manual
Change-Id: Ibc9d5374d44955493a41b477482d5585546c7eb1

Former-commit-id: 05bfd051303f0b500c87c5ba756fa0940836573b