[Bouncer] Remove margin top for bouncer view.
The bouncer view has a slight scroll feedback with touch and clips the children when scrolled up too far.
This is because of a parent container that has another child and has a margin top. I have moved the bouncer view
to the root view. This is fine because the sibling keyguard message area
is now only used when bouncer is not showing. Additionally I took the
sum of those margins and set it as the padding top in the view binder.

Fixes: 260477609
Test: Open bouncer.
Change-Id: Icdf5a314115693d9f7188971adba1396dac105e6

Former-commit-id: 9271f66c370d7b45a3caee051e5c68c2805391ba