[SettingsProvider] IpcDataCache for Settings
(Currently disabled, so this CL will be a no-op, only adding some
implementations.)

The old caching mechanism causes performance janks because the all the
settings of the same type share the same cache. An update of a single
setting leads to the invalidation of the whole cache.

This CL adds a new caching mechanism, using IpcDataCache. In this CL,
all settings of a same type still shares one cache. In the next CL,
clusters of caches will be created based on the frequency of cache
invalidations.

BUG: 228619157
Test: atest SettingsProviderTest
Change-Id: Id823994eebafb4d15f31502ba6a96f70892d8e6d

Former-commit-id: d4e1a667ebbe7a1d939d18e09158bde10f14b0d6