Clean up some rotation utility methods
Most of methods of RotationAnimationUtils were moved to other
class such as TransitionAnimation. One of the remaining method
is duplicated. So move the last method to CoordinateTransforms.

Assume the rotation animation related code in wm core will be
replaced by wm shell in the future, so prepare an initial cleanup.

Bug: 163976519
Test: CoordinateTransformsTest
Change-Id: Ibcf4aad7c771908b81150708db164e4e350935c4

Former-commit-id: 3da3dc28912e09f375675d3b00ca2399b4c7e915