Fix ActivityEmbedding placeholder launching above fullscreen activity
Before, we always place the new created TaskFragment on the top of the
Task. Now, when creating TaskFragment for placeholder, we place the new
TasKFragment right above the primary TaskFragment.

In case there is a fullscreen transparent Activity on top of the primary
Activity, this prevents us from launching the placeholder on top of the
transparent Activity.

Bug: 261550242
Test: atest WmTests:TaskFragmentOrganizerControllerTest
Change-Id: I163e9ca1637db45e8695d40a31f2a2c6a4f05189

Former-commit-id: 7ed3a9c3f9a08623263667cb8775164239f4c833