Optimistic fix for tile not getting redistributed in split shade
The issue is likely caused by QS tiles not getting redistributed in PagedTileLayout#onConfigurationChanged which does it only when orientation changes. One fix would be to remove that condition at all but then we would be redistributing for every configuration change or we'd need to make PagedTileLayout aware of split shade - none of these options seem good.

So the solution here is to force redistribution from QSPanelControllerBase which is already aware of split shade and can listen to changes there.

I'll also add more logs (in case bug is not fixed) in follow-up CL.

Bug: 255208946
Test: there is no reliable way to reproduce the issue, we'll see if it stops happening
Test: QSPanelControllerTest
Change-Id: I81630b6eec16237a1a665982553dfc5b9dbac54b

Former-commit-id: 8b8aa7d66e16d0dc5fd7f25b9b43bcd34e86f3bb