[Bouncer] refine entry for bouncer user switcher.
Match the animation of the user switcher with the pin view.
Increase the duration of the animation and add a y translation animation
when invoking the appear animation.

Also refine the entry text animation. Ensure that prompt is shown first
by removing animation time. Also ensure that initial text does not
animate in so that the entry animation is better refined. Added tests as
well.

For some of the tests, ktfmt reformatted the tests...these should just
be formatting changes.

Bug: 261571180
Test: Manually open bouncer with a quick swipe up in tablet device.
Test: Added unit tests.

Change-Id: Ic36b1554a7c3c73e0e8805db972e1bd3f0de57fe

Former-commit-id: 66430ca2921f899df7fcb6e47da7caec6c809b00