Increase threshold for ESOS power button press detection
If pressing power button 5 times is faster than 200ms, we will
not trigger the ESOS experience. This is unlikely to be done by human. The fastest recorded button press is 60ms so the whole esos gesture sequence should be 300ms or so. We are taking 200ms threshold to be conservative.

Fix: 261618088
Test: rebuild, PTS
Change-Id: Id3f1d92623d17b2921cc0fb90c1f19f896946881

Former-commit-id: ab68e38786ecfa253cca5808182f42b17371db70