Do not specify root view in TouchInsetManager.
TouchInsetManager helps limit the touch targets for a surface to a
given set of child views. Previously, a root view was specified at
construction, from which the attached surface was retrieved from.
Since this information is associated with each tracked view,
providing this information upfront is not necessary.

This changelist updates TouchInsetManager to retrieve the attached
surface from the views rather than at construction. The logic has
been updated to support multiple root surfaces, allowing the same
TouchInsetManager to be used across surfaces.

Test: atest TouchInsetManagertest
Bug: 261781069
Change-Id: Ief02e781fd3b2beb1373f05192690f664cacb6c1

Former-commit-id: bdd33e3b59d95df0ab6e80fa59556a29266c357f