Keep rotation transform for exiting PiP with fixed rotation
The exact fix is the one line change in PipTransition. The example:
The landscape PiP activity is expanding to fullscreen portrait.
So at the end of animation, the task needs to keep the 90 degree
transform. Otherwise it cannot match the area of portrait display.
And then display will continue to update to landscape seamlessly
and the default finish transaction will reset the rotation transform.

The other changes are just small cleanup on the way.

Bug: 260925940
Test: Enable shell transition.
      Use Chrome to play a Youtube video in fullscreen landscape.
      Enter PiP mode by swiping to home.
      Press the PiP window to expand to fullscreen landscape.
      There should be no flickering by inconsistent orientation.
Change-Id: I241eb8fc59af4733794a0e0b4d4438e59edbfe10

Former-commit-id: 56647c4c75915f2dde42823dd86d6e23019efa0f