Support multiple broadcast receivers
Bug: 253906553

Allow system server to schedule multiple broadcast receivers in a
process with a single AIDL call.  The AIDL call takes an array of
receivers.  Each element is either a manifest receiver or a registered
receiver.  This is a do-no-harm commit.  It adds the AIDL call but
does not use it.

The client side processes the list by forwarding the elements to the
scheduleReceiver() and scheduleRegisteredReceiver() handlers, as
appropriate.

Test: atest
 * FrameworksCoreTests

Change-Id: Ib61f8170badcb2fb24a466dc10c607baeeeb4dc3

Former-commit-id: 5a4b0acc74c8ede0a11dae79b89f94f609b77efa