Fix drawing the ring around adaptive app icons
It seems like when you set the bounds on an adaptive icon, it does
not draw the icon in the center of those bounds.

Fix by translating the drawable before setting the bounds.

Also removes drawing black first as this seems to leave a slight
artifact around the edges of the circle otherwise.

Bug: 260636233
Test: manual - mark a bubbleable notification as an important person
             - bubble it
             => observe the app icon badge is centered in the
                highlight ring
Change-Id: I04279a3c51badd26a5141a8edf77f05472cb9718

Former-commit-id: 0c1dd96b1f08c9c27d3edca10865820b3e6968fb