Allow any user's UsageStats to be backed up and restored
Refactor update mappings jobs and prune jobs to use the
user id as the job id and use job namespaces.

Plumb the userId through from the system backup agent

Put 0 in the userId field of the payload to ensure that it
can restore on a single-user device even if backed up from
a secondary user. Henceforth the userId in the payload should
be ignored

Test: Unlock user and dumpsys jobscheduler to look for update job
      Uninstall an app and look for prune job
      Manually force a backup from HSUM device main user and
      restore the dataset on a non-HSUM device

Bug: 260831016

Change-Id: I64904664eb0c949415a9a05eb919674d74081aba

Former-commit-id: 807128ead99b00b03643ad451bb2d3c6f7b61bdf