[Sb refactor] Move WifiActivityModel to shared
Mobile will want to use the same model, so renamed `WifiActivityModel`
to `DataActivityModel` and updated all usages.

The only non-pure part of the change is that `ACTIVITY_PREFIX` changed
not to include the "wifi" part of the name, so I left the constant there
(which will be fine for mobile) but allowed WifiRepositoryImpl to
declare its own prefix for logging.

Test: tests in tests/src/com/android/systemui/pipeline/wifi/*
Bug: 238425913
Change-Id: I4b913414cdacb5e0751e32b4d690685c6ab7d17a

Former-commit-id: d4c86b4e982c939987dffbc241825ce3ea953eb9