Enables flag for the feature of drag to remove and the new physics-based animations ver2.
Goal:
The accessibility floating menu has been refactored and supports the drag to remove, which will be released in Android U. Hopefully let more people experience and test it as possible before releasing it to the next version. Besides, improve and enhance the quality as soon as receiving user feedback and fix it.

Action:
Change the flag from unreleased to released.

Bug: 237716018
Test: manual test
Change-Id: I3f336c7648bab756c2f33e080eae9a43078cde28

Former-commit-id: 6a0182bf8635de6f99ad85facd5d9c97042774e3