Set Visibility on ClockSwitch Animation
This sets the visibility of the clock frames to invisible when they
are removed from view by KeyguardClockSwitch. This does not cause a
visual change with the clock animation directly, but allows code to
use visibility checks of the frames to determine which size clock
is shown. There is an impact on the AOD positioning for the small
clock which was already checking parameters this way, and always
using the large clock positioning parameters as a result. This is
now corrected and KeyguardClockPositionAlgorithm will recieve the
correct parameters depending on whether the large or small clock
is shown.

Fixes: 261755021
Test: Manually checked a few devices
Change-Id: I2bdca38cc70df03dfd952c94559ae8f36e2dc80a

Former-commit-id: f07b624784353c2385dcaa81bd34fc36b331ebb1